fix: refactoring to use both a client- and server-side ServiceLocator
