+++ src/pages/api/admin/bulk-data-import.ts
@@ -2,7 +2,7 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 import { ReasonPhrases, StatusCodes } from "http-status-codes";
 import { ErrorWithCause } from "pony-cause";
-import { services } from "../../../services/ServiceLocator";
+import { services } from "../../../services/ServiceLocatorServer";
 import { BulkDataImportStatus } from "../../../services/AppModel";
 import { serverLogError } from "../log";
