+++ src/pages/api/gateway/[gatewayUID]/node/[nodeId]/config.ts
@@ -3,7 +3,7 @@ import { ErrorWithCause } from "pony-cause";
 import { ReasonPhrases, StatusCodes } from "http-status-codes";
 import type { NextApiRequest, NextApiResponse } from "next";
 
-import { services } from "../../../../../../services/ServiceLocator";
+import { services } from "../../../../../../services/ServiceLocatorServer";
 import { HTTP_STATUS } from "../../../../../../constants/http";
 import { serverLogError } from "../../../../log";
